Start migrating invites endpoints
