+++ server/service/transport_invites.go
@@ -2,24 +2,10 @@ package service
 
 import (
 	"context"
-	"encoding/json"
 	"net/http"
-	"strings"
 
 	"github.com/gorilla/mux"
 )
-
-func decodeCreateInviteRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req createInviteRequest
-	if err := json.NewDecoder(r.Body).Decode(&req.payload); err != nil {
-		return nil, err
-	}
-	if req.payload.Email != nil {
-		*req.payload.Email = strings.ToLower(*req.payload.Email)
-	}
-
-	return req, nil
-}
 
 func decodeDeleteInviteRequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	id, err := uintFromRequest(r, "id")
